backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# CORS and HTTP
python-multipart==0.0.6

# Caching
cachetools==5.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Data validation
python-dateutil==2.8.2

# cTrader integration
ctrader-open-api==0.9.2
twisted==24.3.0
protobuf==3.20.1
python-dotenv==1.0.0
pytz==2024.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0
